requirements.txt
aiogram
faster-whisper
pydub
pandas
openpyxl
google-cloud-aiplatform
python-dotenv
